model updated

+++ Fakturama-Parent/com.sebulli.fakturama.rcp/src/main/java/com/sebulli/fakturama/dao/AbstractDAO.java
@@ -68,7 +68,7 @@ public abstract class AbstractDAO<T> {
      * @param id
      * @return
      */
-    public T findById(int id) {
+    public T findById(long id) {
     	return getEntityManager().find(getEntityClass(), id);
     }
